remove core
