+++ examples/KTO/model_config.py
@@ -15,7 +15,7 @@
 from dataclasses import dataclass, field
 from typing import List, Optional
 
-from core import flatten_dict
+from utils import flatten_dict
 
 
 @dataclass
